add api route and jwt auth

+++ database/seed/user-seeder.go
@@ -22,8 +22,8 @@ func seedUserTable(count int){
 			Email		:	fake.EmailAddress(),
 			Password	:	string(hashedPassword),
 			Gender		:	fake.Gender(),
-			TinggiBadan	:	rand.Intn(180-50) + 50,
-			BeratBadan	:	rand.Intn(200-40) + 40,
+			TinggiBadan	:	float64(rand.Intn(180-50) + 50),
+			BeratBadan	:	float64(rand.Intn(200-40) + 40),
 			Usia		:	rand.Intn(100),
 		})
 	}
